scripts/variantstore/scripts/requirements.txt
google-cloud-bigquery
ijson
numpy
google-cloud-storage
firecloud
terra-notebook-utils
pybedtools
